+++ release_notes.md
@@ -4,6 +4,7 @@
 -->
 - Read FuncitonsHostingConfigurations to enable worker concurrency feature (#8125)
 - Updated Java Worker Version to [1.11.0](https://github.com/Azure/azure-functions-java-worker/releases/tag/1.11.0)
+- Add Include minimum durable payload to sync trigger (#8292)
 
 **Release sprint:** Sprint 118
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+118%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+118%22+label%3Afeature+is%3Aclosed) ]